Moves |read_json_file(file_path)|convert_to_pil(label)| to |hd_utility.py| library module
